+++ mathops/xt_intary_07_test.go
@@ -12,7 +12,7 @@ func TestIntAry_SetIntAryWithInt32_01(t *testing.T) {
 	num := int32(123456789)
 
 	eNumStr := "123456.789"
-	ePrecision := uint(3)
+	ePrecision := 3
 	eSignVal := 1
 
 	ia := IntAry{}.New()
@@ -42,7 +42,7 @@ func TestIntAry_SetIntAryWithInt32_02(t *testing.T) {
 	num := int32(-123456789)
 
 	eNumStr := "-12345.6789"
-	ePrecision := uint(4)
+	ePrecision := 4
 	eSignVal := -1
 
 	ia := IntAry{}.New()
@@ -72,7 +72,7 @@ func TestIntAry_SetIntAryWithInt32_03(t *testing.T) {
 	num := int32(0)
 
 	eNumStr := "0.0000"
-	ePrecision := uint(4)
+	ePrecision := 4
 	eSignVal := 1
 
 	ia := IntAry{}.New()
@@ -102,7 +102,7 @@ func TestIntAry_SetIntAryWithInt32_04(t *testing.T) {
 	num := int32(32)
 
 	eNumStr := "0.0032"
-	ePrecision := uint(4)
+	ePrecision := 4
 	eSignVal := 1
 
 	ia := IntAry{}.New()
@@ -132,7 +132,7 @@ func TestIntAry_SetIntAryWithInt32_05(t *testing.T) {
 	num := int32(-32)
 
 	eNumStr := "-32"
-	ePrecision := uint(0)
+	ePrecision := 0
 	eSignVal := -1
 
 	ia := IntAry{}.New()
@@ -162,7 +162,7 @@ func TestIntAry_SetIntAryWithInt32_06(t *testing.T) {
 	num := int32(32)
 
 	eNumStr := "0.32"
-	ePrecision := uint(2)
+	ePrecision := 2
 	eSignVal := 1
 
 	ia := IntAry{}.New()
@@ -372,7 +372,7 @@ func TestIntAry_SetIntAryWithInt64_01(t *testing.T) {
 	num := int64(123456789)
 
 	eNumStr := "123456.789"
-	ePrecision := uint(3)
+	ePrecision := 3
 	eSignVal := 1
 
 	ia := IntAry{}.New()
@@ -402,7 +402,7 @@ func TestIntAry_SetIntAryWithInt64_02(t *testing.T) {
 	num := int64(-123456789)
 
 	eNumStr := "-12345.6789"
-	ePrecision := uint(4)
+	ePrecision := 4
 	eSignVal := -1
 
 	ia := IntAry{}.New()
@@ -432,7 +432,7 @@ func TestIntAry_SetIntAryWithInt64_03(t *testing.T) {
 	num := int64(0)
 
 	eNumStr := "0.0000"
-	ePrecision := uint(4)
+	ePrecision := 4
 	eSignVal := 1
 
 	ia := IntAry{}.New()
@@ -462,7 +462,7 @@ func TestIntAry_SetIntAryWithInt64_04(t *testing.T) {
 	num := int64(32)
 
 	eNumStr := "0.0032"
-	ePrecision := uint(4)
+	ePrecision := 4
 	eSignVal := 1
 
 	ia := IntAry{}.New()
@@ -492,7 +492,7 @@ func TestIntAry_SetIntAryWithInt64_05(t *testing.T) {
 	num := int64(-32)
 
 	eNumStr := "-32"
-	ePrecision := uint(0)
+	ePrecision := 0
 	eSignVal := -1
 
 	ia := IntAry{}.New()
@@ -522,7 +522,7 @@ func TestIntAry_SetIntAryWithInt64_06(t *testing.T) {
 	num := int64(32)
 
 	eNumStr := "0.32"
-	ePrecision := uint(2)
+	ePrecision := 2
 	eSignVal := 1
 
 	ia := IntAry{}.New()
