+++ mathops/xt_numstrdto_03_test.go
@@ -161,11 +161,11 @@ func TestNumStrDto_GetAbsIntRunes_03(t *testing.T) {
 func TestNumStrDto_GetBigIntNum_01(t *testing.T) {
 
 	bigI := big.NewInt(int64(123456123456))
-	precision := uint(6)
+	precision := 6
 
 	exStr := "123456.123456"
 
-	expectedBigIntNum := BigIntNum{}.NewBigInt(bigI, precision)
+	expectedBigIntNum := BigIntNum{}.NewBigInt(bigI, uint(precision))
 
 	intAry, err := IntAry{}.NewBigInt(bigI, precision)
